refactor: move CoerceStringArray to ToolParams and use GetRaw for snake_case fallback
Remove GetParam from ParamCoercion — ToolParams.GetRaw() already handles
snake_case/camelCase fallback via StringCaseUtility. Move CoerceStringArray
into ToolParams as GetStringArray() to centralize parameter handling.
Update RunTests and GetTestJob to use ToolParams instead.
